Update .gitignore
First, thank you for sharing your work on building a Class-0/Type-0/MWI-sender. This app is amazing and has a lot of potential.

This commit adds some other files from `.idea` (`gradle.xml` and `misc.xml`) to the `.gitignore` list.

Also, I removed `.directory` because I'm not sure if it was there on purpose, so let me know if it shouldn't be removed.

Cheers.